requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.4
django-allauth==0.32.0
django-angular==0.8.3
django-debug-toolbar==1.6
django-debug-toolbar-request-history==0.0.3
django-debug-toolbar-template-profiler==1.0.1
django-debug-toolbar-template-timings==0.7
djangorestframework==3.5.3
idna==2.6
mysqlclient==1.3.10
oauthlib==2.0.2
pbr==3.1.1
python-openid==2.2.5
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
sqlparse==0.2.4
stevedore==1.27.1
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.8.2
